* Fitness function is ok

+++ src/robby.h
@@ -4,7 +4,7 @@
 #include "world.h"
 
 void place_robby(world *w, int col, int row);
-int move_robby(world *w, int col0, int row0, int col1, int row1);
+float move_robby(world *w, int col0, int row0, int col1, int row1);
 void get_robby_position(world *w, int *col, int *row);
 void get_neighbours(world *w, unsigned int *n);
